Simplify the codebase
The previous version suffered strongly from being overbuilt;
there's no reason why the events library had to have multiple
codepaths for different eventloops, or why the select codepath
had to reimplement everything.

Similarly, implementing pipeline on the driver level was just a
plain old bad idea.  Maybe with the X protocol that will change,
but here it just lead to absurdly complicated code for something
that I personally never ended up even considering.

Some benching of a high-performance system using the previous
version showed that almost as much time was spent waiting for
mysql, as we spent doing busywork in the perl code.  That's just
not kosher.

So this commit tries to yank out as much cruft as possible.