Do not use the bundled zlib
When compiling mariadb-connector-c, by default it compiles
a bundled zlib and uses that.

There were some nasty bugs in MySQL's Connector/C due to
exactly this -- the bundled zlib ended up interacting poorly
with the system zlib.

Out of an abundance of caution, disabling the bundled zlib,
thus making this always use the system zlib.